requirements.txt
numpy>=1.17
scipy>=1.4
qiskit-terra>=0.19.0
qiskit-ibmq-provider>=0.16.0
matplotlib>=3.4
uncertainties

